Match optional space inside the GPU model number
We've seen samples of prefixed GPU model numbers given with both a space inside and no space inside. Examples:

Radeon Pro WX3200 vs.
Radeon Pro WX 3200

Change the name matching algorithm so that the version number may contain a short uppercase prefix and so that there's an optional space between the prefix and the digits.